requirements.txt
flask
flask-socketio
discord.py
sqlalchemy